OGM-110 Rename package
