+++ hibernate-ogm-core/src/main/java/org/hibernate/ogm/datastore/mapbased/impl/MapAssociationSnapshot.java
@@ -18,7 +18,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA  02110-1301, USA.
  */
-package org.hibernate.ogm.datastore.impl.mapbased;
+package org.hibernate.ogm.datastore.mapbased.impl;
 
 import java.util.Map;
 import java.util.Set;
