+++ hibernate-ogm-core/src/test/resources/hibernate.properties
@@ -25,4 +25,4 @@
 #@author Emmanuel Bernard
 hibernate.dialect = org.hibernate.ogm.dialect.NoopDialect
 hibernate.search.lucene_version=LUCENE_CURRENT
-hibernate.ogm.datastore.provider=org.hibernate.ogm.datastore.impl.mapbased.MapBasedDatastoreProvider
+hibernate.ogm.datastore.provider=org.hibernate.ogm.datastore.mapbased.impl.MapBasedDatastoreProvider
